requirements-dev.txt
aiohttp==3.8.5
pytest==7.4.0
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-homeassistant-custom-component==0.13.5
pre-commit==3.4.0
black==23.7.0
flake8==6.1.0
isort==5.12.0
mypy==1.5.1
pylint==2.17.5
python-dateutil==2.8.2
voluptuous==0.13.1